% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{a_efficiency}
\alias{a_efficiency}
\title{A-efficiency}
\usage{
a_efficiency(x)
}
\arguments{
\item{x}{A variance-covariance matrix}
}
\value{
A single efficiency measure
}
\description{
Computes the A-efficiency of the design, which is equal to the trace of the
variance-covariance matrix over the number of parameters to be estimated
}
\references{
Bliemer and Rose, 2009, Efficiency and sample size requirements for state choice experiments, Transportation Research Board Annual Meeting, Washington DC
Scarpa and Rose, 2008, Designs efficiency for non-market valuation with choice modelling: How to measure it, what to report and why, Australian Journal of Agricultural and Resource Economics, 52(3):253-282
}
