% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{all_priors_and_levels_specified}
\alias{all_priors_and_levels_specified}
\title{Check whether all priors and attributes have specified levels}
\usage{
all_priors_and_levels_specified(x)
}
\arguments{
\item{x}{A list of utility expressions}
}
\value{
A boolean equal to `TRUE` if all are specified and `FALSE` if not
}
\description{
Check whether all priors and attributes have specified levels
}
