% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{any_duplicates}
\alias{any_duplicates}
\title{Check whether any priors or attributes are specified with a value more than
once}
\usage{
any_duplicates(x)
}
\arguments{
\item{x}{A list of utility expressions}
}
\value{
A boolean equal to `TRUE` if specified more than once.
}
\description{
Check whether any priors or attributes are specified with a value more than
once
}
