% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{apply_restrictions}
\alias{apply_restrictions}
\title{Apply restrictions to the candidate set}
\usage{
apply_restrictions(candidate_set, restrictions)
}
\arguments{
\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no
candidate set is provided, then the full factorial subject to specified
restrictions will be used.}

\item{restrictions}{A list of restrictions. Often this list will be pulled
directly from the list of options or it is a modified list of restrictions
following calls to _dummy or _effects coding.}
}
\value{
A restricted candidate set
}
\description{
The function takes the list of restrictions and transforms them into an
expression that is then parsed and evaluated to apply the restrictions
to the supplied candidate set using standard subsetting routines.
}
