% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_whole_word}
\alias{as_whole_word}
\title{Match names as whole words}
\usage{
as_whole_word(name)
}
\arguments{
\item{name}{A character vector of names}
}
\value{
A character vector of regular expressions
}
\description{
Creates regular expressions that match names as whole words, so that e.g.
x1 does not match x10, b_x1 or alt1_x1. Names in the utility functions only
contain letters, digits and underscores.
}
