% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{attribute_level_balance}
\alias{attribute_level_balance}
\title{Check whether we can achieve attribute level balance}
\usage{
attribute_level_balance(x, rows)
}
\arguments{
\item{x}{A list of utility expressions}

\item{rows}{The number of rows in the design}
}
\value{
A boolean equal to `TRUE` if attribute level balance can be achieved
and `FALSE` otherwise
}
\description{
Check whether we can achieve attribute level balance
}
