% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{attribute_levels}
\alias{attribute_levels}
\title{Generic#' Generic for getting the attributes and levels from the utility}
\usage{
attribute_levels(x)
}
\arguments{
\item{x}{An object of class utility}
}
\description{
Generic#' Generic for getting the attributes and levels from the utility
}
