% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{attribute_names}
\alias{attribute_names}
\title{Generic for getting the attribute names}
\usage{
attribute_names(x)
}
\arguments{
\item{x}{An object of class utility}
}
\value{
A character vector of attribute names
}
\description{
Generic for getting the attribute names
}
