% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_a_efficiency}
\alias{calculate_a_efficiency}
\title{A-efficiency}
\usage{
calculate_a_efficiency(design_vcov)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by \code{\link{derive_vcov}} or
returned by an estimation routine. The matrix should be symmetrical and K-by-K}
}
\value{
A single efficiency measure
}
\description{
Computes the A-efficiency of the design, which is equal to the trace of the
variance-covariance matrix over the number of parameters to be estimated
}
