% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_a_error}
\alias{calculate_a_error}
\title{A-error}
\usage{
calculate_a_error(design_vcov)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by
\code{\link{derive_vcov}} or returned by an estimation routine. The matrix
should be symmetrical and K-by-K}
}
\value{
A single error measure
}
\description{
Computes the A-error of the design, which is equal to the trace of the
variance-covariance matrix over the number of parameters to be estimated
}
