% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_d_efficiency}
\alias{calculate_d_efficiency}
\title{D-efficiency}
\usage{
calculate_d_efficiency(design_vcov)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by \code{\link{derive_vcov}} or
returned by an estimation routine. The matrix should be symmetrical and K-by-K}
}
\value{
A single number
}
\description{
Computes the D-efficiency of the design, which is equal to the K-root of the
determinant of the variance-covariance matrix.
}
