% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_d_error}
\alias{calculate_d_error}
\title{D-error}
\usage{
calculate_d_error(design_vcov)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by
\code{\link{derive_vcov}} or returned by an estimation routine. The matrix
should be symmetrical and K-by-K}
}
\value{
A single number
}
\description{
Computes the D-error of the design, which is equal to the K-root of the
determinant of the variance-covariance matrix.
}
