% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_error_measures}
\alias{calculate_error_measures}
\title{Calculates the error measures}
\usage{
calculate_error_measures(p, design_environment, error_measures_string, opts)
}
\arguments{
\item{p}{A vector of parameters}

\item{design_environment}{A design environment}

\item{error_measures_string}{A string of efficiency criteria to be calculated.
The inputs matches \code{\link{calculate_efficiency_criteria}} type.}

\item{opts}{List of options}
}
\value{
A named vector of error measures
}
\description{
The function is called from inside the lapply() loop over priors in the
\code{\link{generate_design}} function.
}
