% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_s_efficiency}
\alias{calculate_s_efficiency}
\title{S-efficiency}
\usage{
calculate_s_efficiency(design_vcov, p, all, significance)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by \code{\link{derive_vcov}} or
returned by an estimation routine. The matrix should be symmetrical and K-by-K}

\item{p}{A vector of parameters, e.g. the named vector of priors. This is used
for c- and s-efficiency. Default value is NULL}

\item{all}{If `TRUE` return a K or K-1 vector with parameter specific error measures.
Default is `FALSE`.}

\item{significance}{A t-value corresponding to the desired level of significance.
The default is significance at the 5% level with an associated t-value of
1.96.}
}
\value{
A vector giving the "minimum" sample size for each parameter or a
single number with the smallest sample size needed for all parameters to be
theoretically significant.
}
\description{
Calculates a "lower bound" sample size to obtain theoretically significant
parameter estimates under the assumption that the priors are correct.
}
