% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_s_error}
\alias{calculate_s_error}
\title{S-error}
\usage{
calculate_s_error(design_vcov, p, return_all, significance)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by
\code{\link{derive_vcov}} or returned by an estimation routine. The matrix
should be symmetrical and K-by-K}

\item{p}{Prior values}

\item{return_all}{If `TRUE` return a K or K-1 vector with parameter specific error
measures. Default is `FALSE`.}

\item{significance}{A t-value corresponding to the desired level of
significance. The default is significance at the 5% level with an associated
t-value of 1.96.}
}
\value{
A vector giving the "minimum" sample size for each parameter or a
single number with the smallest sample size needed for all parameters to be
theoretically significant.
}
\description{
Calculates a "lower bound" sample size to obtain theoretically significant
parameter estimates under the assumption that the priors are correct.
}
