% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_candidate_set}
\alias{check_candidate_set}
\title{Run checks on the candidate set}
\usage{
check_candidate_set(candidate_set, parsed_utility)
}
\arguments{
\item{candidate_set}{A candidate set}

\item{parsed_utility}{A list with the parsed utility expressions from
\code{\link{parse_utility}}}
}
\value{
Nothing
}
\description{
\code{\link{generate_design}} runs three sets of checks, all detailed here,
in the order of appearance of the function arguments. The checks on the
'candidate_set' is to ensure that the user supplied candidate set is valid.
This mostly comes down to checking its dimensions against the number of
attributes and levels in the utility functions.
}
\details{
If no candidate set is supplied, \code{\link{generate_design}} will make a
call to \code{\link{generate_full_factorial}} to use the full factorial
subject to specified constraints as the candidate set.
}
