% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{clean_utility}
\alias{clean_utility}
\title{Cleans the utility expression}
\usage{
clean_utility(x)
}
\arguments{
\item{x}{An object of class utility}
}
\value{
A cleaned utility function as a list
}
\description{
The function cleans the utility expression by removing extra white spaces,
removes brackets and other information to return a clean, easy-to-read
expression
}
