% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.spdesign}
\alias{coef.spdesign}
\title{Generic for extracting the vector of priors}
\usage{
\method{coef}{spdesign}(object, ...)
}
\arguments{
\item{object}{A model object of class 'spdesign'}

\item{...}{Additional arguments passed to the function}
}
\value{
A vector of named priors used in the optimization
}
\description{
Generic for extracting the vector of priors
}
