% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{contains_dummies}
\alias{contains_dummies}
\title{Check whether the utility function contains dummy coded variables}
\usage{
contains_dummies(string)
}
\arguments{
\item{string}{A string or list of strings}
}
\value{
A boolean equal to `TRUE` if the utility function contains dummy
coded attributes and `FALSE` otherwise
}
\description{
We are splitting on all separators first before detecting whether we have
dummy coded attributes to allow for people reusing the _dummy name for the
attribute.
}
