% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cor}
\alias{cor}
\title{Correlation}
\usage{
cor(x, ...)
}
\arguments{
\item{x}{A model object of class 'spdesign'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate the correlation of the design. This is a wrapper around
\code{\link{cor}}.
}
