% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{cycle}
\alias{cycle}
\title{Cycling of attribute levels}
\usage{
cycle(current_design_candidate)
}
\arguments{
\item{current_design_candidate}{A candidate for the RSC algorithm. To allow
for marginal changes the candidate is passed back in for each iteration.}
}
\value{
A cycled design candidate
}
\description{
Cycles the attribute levels to create a new design candidate. "Cycling
replaces all attribute levels in each choice situation at the tiem by
replacing the first level with the second level, second level with the third
etc. Since this change affects all columns, cycling can only be performed if
all attributes have exactly the same sets of feasible levels,
(e.g., where all variables are dummy coded)." (p. 253).
}
\details{
This part of the RSC algorithm is rarely invoked.
}
\references{
Hensher, D. A., Rose, J. M. & Greene, W., 2005, Applied Choice Analysis,
2nd ed., Cambridge University Press
}
