% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{d_efficiency}
\alias{d_efficiency}
\title{D-efficiency}
\usage{
d_efficiency(x)
}
\arguments{
\item{x}{A variance-covariance matrix}
}
\value{
A single number
}
\description{
Computes the D-efficiency of the design, which is equal to the K-root of the
determinant of the variance-covariance matrix. In the case of Bayesian priors,
the formula is averaged over the draws used to simulate the "prior distribution".
}
\references{
Kessels, R., Goos, P. and Vandebroek, M., 2006, A comparison of criteria to design efficient choice experiments, Journal of Marketing Research, 43(3):409-419
Scarpa and Rose, 2008, Designs efficiency for non-market valuation with choice modelling: How to measure it, what to report and why, Australian Journal of Agricultural and Resource Economics, 52(3):253-282
}
