% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xj.R
\name{define_base_x_j}
\alias{define_base_x_j}
\title{Define base x_j}
\usage{
define_base_x_j(utility, design_candidate)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}

\item{design_candidate}{The current design candidate under consideration}
}
\value{
A base list x_j with model matrices the lenght of J
}
\description{
Defines the base of the x_j list using the parsed utility expression,
design_candidate and the base model matrix
}
