% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_vcov_mnl}
\alias{derive_vcov_mnl}
\title{Derive the variance covariance matrix for the MNL model}
\usage{
derive_vcov_mnl()
}
\value{
The variance co-variance matrix
}
\description{
The function takes no arguments and is evaluated in context!
}
