% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designer.R
\name{design}
\alias{design}
\title{Create the experimental design}
\usage{
design()
}
\value{
A design object of class designer.design
}
\description{
Design takes the utility functions and design options as inputs and generates
candidate designs. Design cannot be run without first validating the utility
functions and design options. Running LINK TO VALIDATE will create a validated
environment, which is the context for evaluating the designs.
}
