% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{digitize}
\alias{digitize}
\title{Expand the sequence of integers}
\usage{
digitize(D, P, count, digit)
}
\arguments{
\item{D}{Number of dimensions}

\item{P}{A vector of prime numbers}

\item{count}{A matrix}

\item{digit}{A vector}
}
\description{
Equation 1 in Bhat (2003)
}
\references{
Bhat, C. R., 2003, Simulation Estimation of Mixed Discrete Choice
Models Using Randomized and Scrambled Halton Sequences, Transportation
Research Part B, 9, pp. 837-855
}
