% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{dummy_names}
\alias{dummy_names}
\title{Find the position of the dummy coded attributes}
\usage{
dummy_names(x)
}
\arguments{
\item{x}{An object of class utility}
}
\value{
A boolean vector matching the expanded utility expression
}
\description{
The function will find the position of the dummy coded attributes in the
candidate set (in the case of the Modified Federov or Random algorithms) or
the design candidate (in the case of the RSC algorithm). This will let us
know which columns to coerce to factors prior to defining x_j.
}
