% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions-and-conditions.R
\name{exclude}
\alias{exclude}
\title{Exclude rows from the candidate set}
\usage{
exclude(candidate_set, exclusions)
}
\arguments{
\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no
candidate set is provided, then it is built from the utility functions
subject to specified exclusions with \code{\link{build_candidate_set}}.
This is passed in as an object and not a character
string. The candidate set will be expanded to include zero columns to
consider alternative specific attributes.}

\item{exclusions}{A list of exclusions Often this list will be pulled
directly from the list of options or it is a modified list of exclusions}
}
\value{
A restricted candidate set
}
\description{
Each exclusion is evaluated with the columns of the candidate set as
variables, and the rows where it is TRUE are removed.
}
