% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{expand_attribute_levels}
\alias{expand_attribute_levels}
\title{Expand the list of attributes and levels to the "wide" format}
\usage{
expand_attribute_levels(x)
}
\arguments{
\item{x}{An object of class utility}
}
\value{
A named vector
}
\description{
Expands the attributes and levels to the wide format. The nested list is
padded with zeros where alternative specific attributes are present to ensure
that we can work with square matrices.
}
