% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_all_names}
\alias{extract_all_names}
\title{Extract all names}
\usage{
extract_all_names(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE.}
}
\value{
A list or vector with all names
}
\description{
Extracts all parameter and attribute names from the utility function.
This is a wrapper around \code{\link[stringr]{str_extract_all}} with a specified
boundary. The function also calls \code{\link{remove_all_brackets}} to
ensure that if a word is used inside a square bracket, e.g. seq, it is not
extracted.
}
\details{
A name is a word that starts with a letter and is not followed by an opening
round bracket. This avoids extracting numbers, e.g. the 2 in I(x1^2), and
functions such as the interaction operator I() as attributes.
}
