% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_attribute_names}
\alias{extract_attribute_names}
\title{Extract attribute names}
\usage{
extract_attribute_names(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE.}
}
\value{
A Vector or string wtih attribute names
}
\description{
Extracts attribute names. It is a wrapper around
\code{\link{extract_all_names}} and \code{\link{extract_param_names}}.
}
