% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_attrs}
\alias{extract_attrs}
\title{Extracts the attributes and levels of the utility functions}
\usage{
extract_attrs(x)
}
\arguments{
\item{x}{A character string}
}
\description{
Extracts the attributes and levels of the utility functions
}
