% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_bayesian_prior}
\alias{extract_bayesian_prior}
\title{Extract Bayesian priors}
\usage{
extract_bayesian_prior(string)
}
\arguments{
\item{string}{A character string}
}
\description{
Extract Bayesian priors
}
