% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_distribution}
\alias{extract_distribution}
\title{Extract distributions}
\usage{
extract_distribution(string)
}
\arguments{
\item{string}{A single character string or list of character strings with a
single or multiple utility functions}
}
\value{
A list of length equal to the number of distributions in the order
that they appear in the utility functions going from top left to bottom right
across the list.
}
\description{
This function will locate and extract the the distributions for Bayesian
priors and random parameters as specified in the design. The output is used
to create the matrix of correct draws for priors and parameters.
}
