% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_dummy}
\alias{extract_dummy}
\title{Extract dummy coded}
\usage{
extract_dummy(string)
}
\arguments{
\item{string}{A character string}
}
\value{
A list with the original input string, a named parameter vector of priors,
and a list of dummy coded attributes and their levels.
}
\description{
The function extracts and returns parameters attributes and levels using the
'.dummy' syntax.
}
