% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_name_args}
\alias{extract_name_args}
\title{Extract the name argument(s)}
\usage{
extract_name_args(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a character vector}
}
\value{
A list
}
\description{
Extract the name argument(s) of the supplied string.
}
