% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_named_values}
\alias{extract_named_values}
\title{Extracts the named values of the utility function}
\usage{
extract_named_values(string)
}
\arguments{
\item{string}{A character string}
}
\value{
A named list of parameter and attribute values. Each list element is
named and can contain a single prior, a list with a mean and sd, or a vector
with attribute levels
}
\description{
The function extracts the named values of the supplied utility function.
}
