% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_names}
\alias{extract_names}
\title{Extract names}
\usage{
extract_names(x)
}
\arguments{
\item{x}{A character vector of the form name | parameter(s)}
}
\value{
A character vector. In most cases the character vector will be of
length 1, but in the case of interaction terms, the character vector will
be of length equal to the number of terms.
}
\description{
Extracts the name of a parameter or attribute by removing everything after
and including |.
}
\examples{
x <- "x_1|c(0,1)"
extract_names(x)

x <- c("x_2|c(2,4,6,8)", "x_3|c(2,5)")
extract_names(x)

}
