% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_numeric}
\alias{extract_numeric}
\title{Extract numeric}
\usage{
extract_numeric(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE}
}
\value{
Returns a list containing all the numerical values in the supplied
vector. Returns a single vector if simplify = TRUE
}
\description{
Extracts the numerical values from a string. The function is a wrapper around
\code{\link{str_extract_all}}.The function is only intended for internal use
to avoid code-duplication.
}
