% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_parsers.R
\name{extract_param}
\alias{extract_param}
\title{Extract the parameters}
\usage{
extract_param(V, bayesian)
}
\arguments{
\item{V}{A list of utility functions}

\item{bayesian}{If TRUE use Bayesian priors for the parameters}
}
\value{
A named vector of parameters
}
\description{
Extracts the parameters from the list of matrices created by
}
