% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_param_distribution}
\alias{extract_param_distribution}
\title{Extract the parameter distribution}
\usage{
extract_param_distribution(string)
}
\arguments{
\item{string}{A single character string or list of character strings with a
single or multiple utility functions}
}
\description{
Extract the parameter distribution
}
