% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_param_names}
\alias{extract_param_names}
\title{Extract parameter names}
\usage{
extract_param_names(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE.}
}
\value{
A list or vector with the parameter names.
}
\description{
Extracts all words starting with "b_". Leverages the fact that all parameters
has to start with "b_".
}
