% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_params}
\alias{extract_params}
\title{Extracts the parameters}
\usage{
extract_params(string)

extract_params(string)
}
\arguments{
\item{string}{A character string}

\item{x}{A character string}
}
\value{
A named vector of parameters with associated attributes for random
parameters and distributions
}
\description{
Wrapping around extract values and extract names
}
