% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_prior_distribution}
\alias{extract_prior_distribution}
\title{Extract the prior distribution}
\usage{
extract_prior_distribution(string)
}
\arguments{
\item{string}{A single character string or list of character strings with a
single or multiple utility functions}
}
\description{
Extract the prior distribution
}
