% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_specified}
\alias{extract_specified}
\title{Extract specified}
\usage{
extract_specified(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE.}
}
\description{
Only extract parameters and attributes with specified priors and levels. This
is very useful to test whether parameters or attributes are specified
multiple times
}
