% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_unparsed_values}
\alias{extract_unparsed_values}
\title{Extract unparsed named values of the utilitiy function}
\usage{
extract_unparsed_values(string)
}
\arguments{
\item{string}{A character string}
}
\value{
A named list of parameter and attribute values. Each list element is
  named and contains a numeric value or expression to be parsed
}
\description{
If the utility function contains parameters that are dummy coded, the dummy
coding is handled here. By expanding the dummy coding prior to parsing we can
directly consider Bayesian priors for each level.
}
