% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_utility_components}
\alias{extract_utility_components}
\title{Extracts the utility components}
\usage{
extract_utility_components(string)
}
\arguments{
\item{string}{A character string formatted as a utility function}
}
\value{
A named list with the utility expression, the trimmed utility without
attribute and parameter information, a named vector with priors and a named
list with attributes and levels.
}
\description{
This function acts as a wrapper around \code{\link{extract_utility_function}},
\code{\link{extract_params}} and \code{\link{extract_attrs}}, and will parse
a single utility function. The function is only meant for internal use and is
called from within \code{\link{parse_utility}}.
}
