% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_utility_function}
\alias{extract_utility_function}
\title{Extract utility function}
\usage{
extract_utility_function(string)
}
\arguments{
\item{string}{A character string}
}
\description{
Extract utility function
}
