% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_value_args}
\alias{extract_value_args}
\title{Extract the value argument(s)}
\usage{
extract_value_args(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE.}
}
\value{
A list
}
\description{
Extracts the value argument(s) of the supplied string. The value argument
is defined as the characters between | and '*', '+' and $ in the supplied
string.
}
