% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_values}
\alias{extract_values}
\title{Extract the values}
\usage{
extract_values(x)
}
\arguments{
\item{x}{A character vector of the form name | parameter(s)}
}
\description{
Extract the values
}
