% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-factorial.R
\name{full_factorial}
\alias{full_factorial}
\title{Generate the full factorial}
\usage{
full_factorial(attrs)
}
\arguments{
\item{attrs}{A named list of attributes and their levels}
}
\value{
A data frame containing the full factorial
}
\description{
This function is deprecated and will be removed in a future version. Use
\code{\link{build_candidate_set}} to build a candidate set from the utility
functions, or \code{\link{expand.grid}} to create the full factorial of a
list of attributes with levels that are different from those in the utility
functions.
}
\details{
The function is a wrapper around \code{\link{expand.grid}} and generates the
full factorial given the supplied attributes.
}
\examples{
attrs <- list(
  a1 = 1:5,
  a2 = c(0, 1)
)

# Instead of full_factorial(attrs)
expand.grid(attrs)

}
