% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc.R
\name{generate_rsc_candidate}
\alias{generate_rsc_candidate}
\title{Generates a candidate for the RSC algorithm}
\usage{
generate_rsc_candidate(utility, rows)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}

\item{rows}{An integer giving the number of rows in the final design}
}
\value{
A matrix with rows equal to the number of choice tasks and columns
equal to the number of attributes in the 'wide' format
}
\description{
The function will populate the design candidate matrix used for the RSC
algorithm.
}
\details{
To ensure attribute level balance or that the number of times an attribute
level occurs is within the range specified in the list level_occurrence,
we sample from the list of occurrences subject to the constraint that the
sum has to be equal to the number of rows, i.e. choice tasks. While this is
a theoretical bottle neck for very large designs, it is still sufficiently
fast. It is only run once every 10 000 iterations of the RSC algorithm.
}
