% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_attribute_level_occurrence}
\alias{get_attribute_level_occurrence}
\title{Get the attribute level occurrence}
\usage{
get_attribute_level_occurrence(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A list with the attribute level occurrences
}
\description{
Get the attribute level occurrence
}
