% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED-get.R
\name{get_attribute_levels}
\alias{get_attribute_levels}
\title{Get the attribute levels}
\usage{
get_attribute_levels(object, ...)
}
\value{
A list with the attribute levels
}
\description{
Get the attribute levels
}
