% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_prior_values}
\alias{get_prior_values}
\title{Get the prior values}
\usage{
get_prior_values(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A list with the prior values
}
\description{
Get the prior values
}
