% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED-get.R
\name{get_utility}
\alias{get_utility}
\title{Get the cleaned utility functions}
\usage{
get_utility(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A list of cleaned utility expressions
}
\description{
Get the cleaned utility functions
}
