% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_utility_clean}
\alias{get_utility_clean}
\title{Get the cleaned utility functions}
\usage{
get_utility_clean(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A list of cleaned utility expressions
}
\description{
Get the cleaned utility functions
}
