% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_utility_formula}
\alias{get_utility_formula}
\title{Get the utility formula}
\usage{
get_utility_formula(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A list with the utility formula
}
\description{
Get the utility formula
}
