% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_bayesian_prior}
\alias{has_bayesian_prior}
\title{Tests whether the utility expression contains Bayesian priors}
\usage{
has_bayesian_prior(v)
}
\arguments{
\item{v}{A string or list of strings}
}
\value{
A boolean
}
\description{
This is useful to create a boolean for flow-control inside \code{\link{generate_design}}
}
