% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_random_parameter}
\alias{has_random_parameter}
\title{Tests whether the utility expression contains random parameters}
\usage{
has_random_parameter(string)
}
\arguments{
\item{string}{A string or list of strings}
}
\value{
A boolean
}
\description{
This is useful to create a boolean for flow-control inside
\code{\link{generate_design}}
}
