% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_balanced}
\alias{is_balanced}
\title{Tests whether a utility function is balanced}
\usage{
is_balanced(string, open, close)
}
\arguments{
\item{string}{A character string}

\item{open}{An opening bracket ( [ or <}

\item{close}{A closing bracket ) ] or >}
}
\value{
A boolean equal to `TRUE` if the utility expression is balanced
}
\description{
Tests whether there is an equal number of opening and closing brackets in
the utility functions.
}
