% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level-balance.R
\name{level_balance}
\alias{level_balance}
\title{Print level balance of your design}
\usage{
level_balance(design, block = FALSE)
}
\arguments{
\item{design}{An spdesign object}

\item{block}{A boolean equal to TRUE if you want frequency tables per block.
The default value is FALSE}
}
\description{
Prints a table of level balance for your design. If the design is blocked
you will get both level balance per block and overall level balance
}
