% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{level_occurrences_specified}
\alias{level_occurrences_specified}
\title{Test whether level occurrences are specified in the utility functions}
\usage{
level_occurrences_specified(utility)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}
}
\value{
A boolean equal to TRUE if level occurrences are specified
and FALSE otherwise
}
\description{
Test whether level occurrences are specified in the utility functions
}
