% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{lvl_occurrences}
\alias{lvl_occurrences}
\title{Attribute level occurrence lookup tables}
\usage{
lvl_occurrences(utility, rows, level_balance)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}

\item{rows}{An integer giving the number of rows in the final design}

\item{level_balance}{Boolean equal to TRUE if level balance. This is not used}
}
\value{
A list the length of the expanced attribute levels. Each list element
is a lookup table where the names of the table is the attribute level and the
element the number of times the minimum number of times the level occurs.
}
\description{
Creates a list of lookup tables for attribute level occurrence.
}
