% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-algorithms.R
\name{make_design_candidate}
\alias{make_design_candidate}
\title{Make a design candidate}
\usage{
make_design_candidate(candidate_set, opts, J, type)
}
\arguments{
\item{candidate_set}{A valid candidate set}

\item{opts}{A list of design options}

\item{J}{Number of alternatives per choice task}

\item{type}{One of the implemented types of optimization algorithms}
}
\value{
A matrix in wide format containing the design candidate
}
\description{
The function is a call to \code{\link{switch}} and a wrapper around the
design algorithms \code{\link{random}}, \code{\link{federov}} and \code{\link{rsc}}.
}
