% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_draws}
\alias{make_draws}
\title{Make random draws}
\usage{
make_draws(N, R, D, seed, type)
}
\arguments{
\item{N}{Number of respondents in your sample}

\item{R}{Number of draws per respondent}

\item{D}{Number of dimensions}

\item{seed}{A seed to change the scrambling of the sobol sequence.}

\item{type}{A character string}
}
\value{
A matrix of dimensions N*R x D of standard uniform draws
}
\description{
A common interface to creating a variety of random draws used to simulate
the log likelihood function
}
\examples{
N <- 10
R <- 5
D <- 3

draws <- make_draws(N, R, D, seed = 10, "scrambled-sobol")
head(draws)

draws <- make_draws(N, R, D, seed = 10, "scrambled-halton")
head(draws)

}
