% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_mlhs}
\alias{make_mlhs}
\title{Make Modified Latin Hypercube Draws}
\usage{
make_mlhs(N, R, D)
}
\arguments{
\item{N}{Number of respondents in your sample}

\item{R}{Number of draws per respondent}

\item{D}{Number of dimensions}
}
\description{
Make Modified Latin Hypercube Draws
}
\references{
Hess, S., Train, K. E. & Polak, J. W., 2006, On the use of a Modified Latin
Hypercube Sampling (MLHS) method in the estimation of a Mixed Logit Model
for vehicle choice, Transportation Research Part B, 40, pp. 147-163
}
