% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_pseudo_random}
\alias{make_pseudo_random}
\title{Make pseudo random draws}
\usage{
make_pseudo_random(N, R, D)
}
\arguments{
\item{N}{Number of respondents in your sample}

\item{R}{Number of draws per respondent}

\item{D}{Number of dimensions}
}
\description{
Wrapper for runif to create a common interface
}
