% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_random_draws}
\alias{make_random_draws}
\title{Make random draws}
\usage{
make_random_draws(N, R, D, type)
}
\arguments{
\item{N}{Number of respondents in your sample}

\item{R}{Number of draws per respondent}

\item{D}{Number of dimensions}

\item{type}{A character string}
}
\description{
A common interface to creating a variety of random draws used to simulate
the log likelihood function
}
\examples{
N <- 10
R <- 5
D <- 3

draws <- make_random_draws(N, R, D, "scrambled_sobol")
head(draws)

draws <- make_random_draws(N, R, D, "scrambled_halton")
head(draws)

}
