% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_scrambled_sobol}
\alias{make_scrambled_sobol}
\title{Make scrambled sobol draws}
\usage{
make_scrambled_sobol(N, R, D)
}
\arguments{
\item{N}{Number of respondents in your sample}

\item{R}{Number of draws per respondent}

\item{D}{Number of dimensions}
}
\description{
Wrapper function for sobol() from randtoolbox to create a common interface.
Owen + Fazure_Tezuka Scrambling
}
