% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_standard_halton}
\alias{make_standard_halton}
\title{Wrapper for halton()}
\usage{
make_standard_halton(N, R, D)
}
\arguments{
\item{N}{Number of respondents in your sample}

\item{R}{Number of draws per respondent}

\item{D}{Number of dimensions}
}
\description{
Wrapper function for halton() from randtoolbox to create a common interface
}
