% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_standard_sobol}
\alias{make_standard_sobol}
\title{Make sobol draws}
\usage{
make_standard_sobol(n_ind, n_draws, n_dim, seed = seed)
}
\arguments{
\item{n_ind}{Number of individuals in your sample}

\item{n_draws}{Number of draws per respondent}

\item{n_dim}{Number of dimensions}

\item{seed}{A seed to change the scrambling of the sobol sequence.}
}
\description{
Wrapper function for sobol() from randtoolbox to create a common interface
}
