% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{min_lvl_occurrence}
\alias{min_lvl_occurrence}
\title{Find minimum level occurrences}
\usage{
min_lvl_occurrence(x, rows)
}
\arguments{
\item{x}{An object of class 'utility' or 'spdesign'}

\item{rows}{Number of rows in the design}
}
\value{
A list of minimum level occurrences for the attribute levels
}
\description{
Find minimium level occurrences. This is useful to ensure/approximate attribute
level balance in designs using the Modified Federov Algorithm or the
Random design algorithms.
}
