% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{N}
\alias{N}
\alias{Np}
\alias{LN}
\alias{LNp}
\alias{TR}
\alias{TRp}
\alias{U}
\alias{Up}
\title{Evaluating a distribution}
\usage{
N(mu, sigma)

Np(mu, sigma)

LN(mu, sigma)

LNp(mu, sigma)

TR(mu, sigma)

TRp(mu, sigma)

U(mu, sigma)

Up(mu, sigma)
}
\arguments{
\item{mu}{A parameter indicating the mean or location of the distribution
depending on whether it is a normal, log-normal, triangular or uniform,
or it can be another call to \code{\link{N}}, \code{\link{LN}}, \code{\link{U}}
or \code{\link{TR}} if the model is an RPL with a Bayesian prior.}

\item{sigma}{A parameter indicating the SD or spread of the distribution
or it can be another call to \code{\link{N}}, \code{\link{LN}}, \code{\link{U}}
or \code{\link{TR}}.}
}
\value{
A list of parameters
}
\description{
The function returns its arguments as a named list. The function is used
inside the utility functions. It is transformed to an expression using
\code{\link{parse}} and evaluated using \code{\link{eval}}. This ensures
that in the case of an RPL with Bayesian priors, recursion is handled
automatically. This significantly simplifies translating the utility function
to lists of parameters to use when optimizing the designs. It is also less
error prone.
}
\section{Functions}{
\itemize{
\item \code{N}: The normal distribution

\item \code{Np}: The normal distribution when applied to a prior

\item \code{LN}: The log normal distribution

\item \code{LNp}: The log-normal distribution when applied to a prior

\item \code{TR}: The triangular distribution

\item \code{TRp}: The triangular distribution when applied to a prior

\item \code{U}: The uniform distribution

\item \code{Up}: The uniform distribution when applied to a prior
}}

\examples{
\dontrun{
eval(parse(text = "N(0, 1)"))
eval(parse(text = "LN(0, 1)"))
eval(parse(text = "TR(0, 1)"))
eval(parse(text = "U(0, 1)"))
}
}
