% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{nlvls}
\alias{nlvls}
\title{Find the number of levels}
\usage{
nlvls(x)
}
\arguments{
\item{x}{An object of class 'utility' or 'spdesign'}
}
\value{
A list with the number of levels for each attribute
}
\description{
Find the number of levels for each attribute
}
