% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{occurrences}
\alias{occurrences}
\title{Generic for extracting the attribute occurrence}
\usage{
occurrences(x, rows)
}
\arguments{
\item{x}{An object of class 'utility' or 'spdesign'}

\item{rows}{Number of rows in the design}
}
\value{
A named list of lists where the outer list is for the attributes and
the inner list, the levels of each attribute and the number or range of times
they can occur
}
\description{
The function gets the list of attributes, levels and occurrences within the
design. This is necessary to impose attribute level balance or near attribute
level balance. This is also necessary to correctly populate the candidate
matrix used for the RSC algorithm.
}
