% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_utility}
\alias{parse_utility}
\title{Parse the utility functions}
\usage{
parse_utility(V)
}
\arguments{
\item{V}{A list of utility functions}
}
\value{
An object containing the outputs from the parsed utility functions, such
as the named vector of priors, the named list of attributes and their levels,
restrictions implied by the utility specifications
}
\description{
The function parses the list of utilities to extract the relevant parameters
used to create the design.
}
