% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prepare_priors}
\alias{prepare_priors}
\title{Prepare the list of priors}
\usage{
prepare_priors(utility, parsed_utility, opts)
}
\arguments{
\item{utility}{A list of utility functions}

\item{parsed_utility}{A list containing the parsed utility expression}

\item{opts}{A list of options}
}
\value{
A list of priors
}
\description{
Prepare the list of priors
}
