% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.utility}
\alias{print.utility}
\title{S3 Generic for printing the utility object}
\usage{
\method{print}{utility}(x, ...)
}
\arguments{
\item{x}{A 'utility' object}

\item{...}{Other parameters passed to print}
}
\description{
S3 Generic for printing the utility object
}
