% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{print_efficiency_criteria}
\alias{print_efficiency_criteria}
\title{Creates a printable version of the efficiency criteria}
\usage{
print_efficiency_criteria(value, criteria, digits = 4, opts = NULL)
}
\arguments{
\item{value}{The value of the efficiency criteria obtained by
\code{\link{calculate_efficiency_criteria}}}

\item{criteria}{A character string with the name of the efficiency criteria.
See manual for valid values}

\item{digits}{The nubmer of digits to round the printed value to. The default
is 4.}

\item{opts}{The list of design options. The default is NULL, but must be
specified to print the criteria in colour.}
}
\value{
A character string.
}
\description{
The function is only meant for internal use to handle pretty printing to
console.
}
