% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-utility.R
\name{priors}
\alias{priors}
\title{Generic for extracting the vector of priors}
\usage{
priors(x)
}
\arguments{
\item{x}{An object of class 'utility' or 'spdesign'}

\item{...}{Additional parameters passed to the function. These are really
only used internally on objects of class 'utility'ß}
}
\value{
A list of named priors used in the optimization
}
\description{
Generic for extracting the vector of priors
}
