% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{probabilities}
\alias{probabilities}
\title{Calculate the probabilities of the design}
\usage{
probabilities(x)
}
\arguments{
\item{x}{An 'spdesign' object.}
}
\value{
A matrix of probabilities for each alternative and choice task.
}
\description{
Will take the design object and calculate the probabilities of each
alternative and choice tasks.
}
\details{
Using Bayesian priors the average across the prior distribution will be used.

Using the specific type of model, either the MNL or RPL probs will be
returned.
}
