% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{probabilities_mnl}
\alias{probabilities_mnl}
\title{Calculate the MNL probabilities}
\usage{
probabilities_mnl(x)
}
\arguments{
\item{x}{An 'spdesign' object.}
}
\value{
A matrix of probabilities for each alternative and choice task. With
Bayesian priors the return is the average probabilites over the prior
distribution
}
\description{
Calculate the MNL probabilities
}
