% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{radical_inverse}
\alias{radical_inverse}
\title{Compute the radical inverse}
\usage{
radical_inverse(n_dim, primes, count, digit, perms)
}
\arguments{
\item{n_dim}{Number of dimensions}

\item{primes}{A vector of prime numbers}

\item{count}{A matrix}

\item{digit}{A vector}

\item{perms}{A matrix of the permutations. Defaults to a set of
Braaten-Weller permutations.}
}
\description{
Equation 2 in Bhat (2003)
}
\references{
Bhat, C. n_draws., 2003, Simulation Estimation of Mixed Descrete
Choice Models Using Randomized and Scrambled Halton Sequences, Transportation
Research Part B, 9, pp. 837-855
}
