% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_design_candidate}
\alias{random_design_candidate}
\title{Create a random design_object candidate}
\usage{
random_design_candidate(
  utility,
  candidate_set,
  rows,
  sample_with_replacement,
  print_counter = FALSE,
  max_attempts = 1e+05
)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}

\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no
candidate set is provided, then it is built from the utility functions
subject to specified exclusions with \code{\link{build_candidate_set}}.
This is passed in as an object and not a character
string. The candidate set will be expanded to include zero columns to
consider alternative specific attributes.}

\item{rows}{An integer giving the number of rows in the final design}

\item{sample_with_replacement}{A boolean equal to TRUE if we sample from the
candidate set with replacement. The default is FALSE}

\item{print_counter}{A boolean equal to TRUE if we print the number of
attempts to find a design candidate every 1000th attempt. The default is
FALSE}

\item{max_attempts}{The maximum number of swaps to try before stopping with
an error. The default is 100000.}
}
\description{
Sample from the candidate set to create a random design_object. If level
occurrences are specified, single rows are swapped with random rows from the
candidate set until the design_object candidate satisfies them. A swap is
kept if it does not move the design_object candidate further away from the
level occurrences.
}
