% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{relabel}
\alias{relabel}
\title{Relabeling of attribute levels}
\usage{
relabel(current_design_candidate)
}
\arguments{
\item{current_design_candidate}{A candidate for the RSC algorithm. To allow for marginal
changes the candidate is passed back in for each iteration.}
}
\description{
Relabels the attribute levels to create a new design candidate. For example,
if the column contains the levels (1, 2, 1, 3, 2, 3) and 1 and 3 are
relabeled, then the column becomes (3, 2, 3, 1, 2, 1), i.e. 1 becomes 3 and
3 becomes 1.
}
\details{
The relabeling is done independently for each column, which implies that
the same attribute will be relabeled differently depending on which
alternative it belongs to.
}
\references{
Hensher, D. A., Rose, J. M. & Greene, W., 2005, Applied Choice Analysis,
2nd ed., Cambridge University Press
}
