% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_param}
\alias{remove_param}
\title{Removes the parameter from the utility string}
\usage{
remove_param(param, string)
}
\arguments{
\item{param}{A string with the parameter name}

\item{string}{A string to remove param from}
}
\description{
Removes the parameter from the utility string
}
