% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_prior}
\alias{remove_prior}
\title{Removes the parameter from the utility string}
\usage{
remove_prior(prior, string)
}
\arguments{
\item{prior}{A string with the parameter name}

\item{string}{A string to remove param from}
}
\description{
Removes the parameter from the utility string
}
