% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_round_brackets}
\alias{remove_round_brackets}
\title{Remove round bracket}
\usage{
remove_round_brackets(string)
}
\arguments{
\item{string}{A character string}
}
\description{
Removes everything between (and including) round brackets. We negating
matches with I(), since this is R's interaction operator.
}
\details{
(?<!I) - A negative lookbehind for I
}
