% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_square_bracket}
\alias{remove_square_bracket}
\title{Removes the square bracket containing parameters and levels}
\usage{
remove_square_bracket(string)
}
\arguments{
\item{string}{A character string}
}
\value{
A string
}
\description{
Takes a utility function (string) as an input and removes all prior and level
specifications. Effectively everything between [] including [].
}
