% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_square_brackets}
\alias{remove_square_brackets}
\title{Remove square bracket}
\usage{
remove_square_brackets(string)
}
\arguments{
\item{string}{A character string}
}
\description{
Removes everything between (and including) square brackets
}
