% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{Remove all white spaces}
\usage{
remove_whitespace(x)
}
\arguments{
\item{x}{A character string}
}
\value{
A character vector with no white spaces
}
\description{
Takes a string as an input and removes all whitespaces in the string
}
\examples{
x <- "  test | this    "
remove_whitespace(x)

}
