% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_cols}
\alias{rep_cols}
\title{Repeat columns}
\usage{
rep_cols(x, times)
}
\arguments{
\item{x}{A matrix or data frame}

\item{times}{An integer indicating the number of times to repeat the
row/column}
}
\value{
A matrix or data.frame depending on the type of the input
}
\description{
Repeats each column of the matrix or data frame 'x' a number of times equal
to 'times'.
}
\examples{
test_matrix <- matrix(runif(12), 4)
rep_cols(test_matrix, 2)

}
