% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_rows}
\alias{rep_rows}
\title{Repeat rows}
\usage{
rep_rows(x, times)
}
\arguments{
\item{x}{A matrix or data frame}

\item{times}{An integer indicating the number of times to repeat the
row/column}
}
\value{
A matrix or data.frame depending on type of the input
}
\description{
Repeats each row in the matrix or data frame 'x' a number of times equal to
'times'.
}
\examples{
test_matrix <- matrix(runif(12), 4)
rep_rows(test_matrix, 2)

}
