% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{s_efficiency}
\alias{s_efficiency}
\title{S-efficiency}
\usage{
s_efficiency(p, x, significance = 1.96, all = FALSE)
}
\arguments{
\item{p}{A vector of parameters}

\item{x}{A variance-covariance matrix}

\item{significance}{A t-value corresponding to the desired level of significance.
The default is significance at the 5% level with an associated t-value of
1.96.}

\item{all}{If `all = TRUE` return a vector with sample size for each parameter,
else return a single number with the minimum theoretical sample size. The
default is FALSE. If used for optimization, then a single number is used, but
the suggested sample size for each parameter is always printed regardless of
efficiency criteria used in optimization.}
}
\value{
A vector giving the "minimum" sample size for each parameter or a
single number with the smallest sample size needed for all parameters to be
theoretically significant. This is equal to the maximum of the vector
}
\description{
Calculates a "lower bound" sample size to obtain theoretically significant
parameter estimates under the assumption that the priors are correct. In the
case of zero priors the formula is undefined and will return Inf.
}
\references{
Kessels, R., Goos, P. and Vandebroek, M., 2006, A comparison of criteria to design efficient choice experiments, Journal of Marketing Research, 43(3):409-419
Bliemer and Rose, 2005a, Efficiency and sample size requirements for stated choice experiments, Report ITLS-WP-05-08, Institute for Transport and Logistics Studies, University of Sydney
Bliemer and Rose, 2009, Efficiency and sample size requirements for stated choice experiments, Transportation Research Board Annual Meeting, Washington DC
}
