% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-defaults.R
\name{set_defaults}
\alias{set_defaults}
\title{Validate design opt}
\usage{
set_defaults(opts_input)
}
\arguments{
\item{opts_input}{A list of user supplied design options}
}
\value{
A list of design options populated by sensible default values
}
\description{
The function takes the list of design options and adds default values where
none are specified. This function is exported, but is not intended to be
called by the user of the package. The function is called from within
\code{\link{design}} to populate the list with sensible defaults
}
