% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{spdesign.coef}
\alias{spdesign.coef}
\title{Generic for extracting the vector of priors}
\usage{
spdesign.coef(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A vector of named priors used in the optimization
}
\description{
Generic for extracting the vector of priors
}
