% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdesign.R
\docType{package}
\name{spdesign}
\alias{spdesign}
\title{spdesign: Designing Stated Preference Experiments in R}
\description{
Contemporary software commonly used to design stated preference experiments
are expensive and the code is closed source. spdesign is a free software
package with an easy to use interface to make flexible stated preference
experimental designs using state-of-the-art methods.
}
