% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdesignR.R
\docType{package}
\name{spdesignR}
\alias{spdesignR}
\alias{spdesignR-package}
\title{spdesignR: Stated Preference Designs in R}
\description{
spdesignR makes it easy to create designs for stated preference experiments.
}
