% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{split_string}
\alias{split_string}
\title{Split a string on the supplied pattern}
\usage{
split_string(string, pattern, simplify = TRUE)
}
\arguments{
\item{string}{A character string to be split}

\item{pattern}{A character string pattern. The pattern should conform to
perl style regex. See \code{\link{strsplit}} for details.}

\item{simplify}{A boolean indicating whether the output should be simplified
from a list to a vector. The default is FALSE and a list of character vectors
is returned. This is the default behavior for \code{\link{strsplit}} and the
default for this wrapper function.}
}
\value{
A list of character vectors.
}
\description{
A simple wrapper around \code{\link{strsplit}}.
}
