% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_parsers.R
\name{split_utility}
\alias{split_utility}
\title{Split utility}
\usage{
split_utility(V)
}
\arguments{
\item{V}{A list of utility functions}
}
\value{
A list of matrices where each list element refers to an alternative.
The matrix has rows equal to the number of parameters for each alternative
and columns equal to param and attributes/interation terms
}
\description{
Takes the list of utility functions and splits them on + and *.
}
