% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_values}
\alias{swap_values}
\title{Swaps the values}
\usage{
swap_values(x, values)
}
\arguments{
\item{x}{A vector}

\item{values}{A vector of length 2 with levels to swap}
}
\value{
A vector of the same length as x with swapped values
}
\description{
Swaps the first value in values with the second value in values in x.
}
