% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{transform_distribution}
\alias{transform_distribution}
\title{Transform distribution}
\usage{
transform_distribution(mu, sigma, eta, type)
}
\arguments{
\item{mu}{A value for the mean of the distribution}

\item{sigma}{A value for the standard deviation of the distribution}

\item{eta}{A numeric standard uniform vector}

\item{type}{The type of distribution}
}
\value{
A vector with the transformed distribution given the parameters
}
\description{
Transform distribution
}
