% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{transform_ln}
\alias{transform_ln}
\title{Transform to the lognormal distribution}
\usage{
transform_ln(mu, sigma, eta)
}
\arguments{
\item{mu}{A value for the mean of the distribution}

\item{sigma}{A value for the standard deviation of the distribution}

\item{eta}{A numeric standard uniform vector}
}
\description{
Transform to the lognormal distribution
}
