% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{transform_n}
\alias{transform_n}
\title{Transform to the normal distribution}
\usage{
transform_n(x)
}
\arguments{
\item{x}{A numeric standard uniform vector}
}
\description{
Transform to the normal distribution
}
