% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{transform_uniform}
\alias{transform_uniform}
\title{Transform to the uniform distribution}
\usage{
transform_uniform(mu, sigma, eta)
}
\arguments{
\item{mu}{A value for the mean of the distribution}

\item{sigma}{A value for the standard deviation of the distribution}

\item{eta}{A numeric standard uniform vector}
}
\description{
Transform to the uniform distribution
}
