% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{unlisted_levels}
\alias{unlisted_levels}
\title{Find attributes with level occurrences and levels that are not listed}
\usage{
unlisted_levels(utility, candidate_set, rows)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}

\item{candidate_set}{A candidate set in wide format}

\item{rows}{Number of rows in the design}
}
\value{
A character vector with the names of the attributes with level
occurrences specified, where the candidate set contains levels that are not
listed in the utility functions. Empty if there are none.
}
\description{
A supplied candidate set may contain attribute levels that are not listed in
the utility functions. This is only a problem for attributes with level
occurrences specified, because occurrences can only be counted for listed
levels.
}
