% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{utility_formula}
\alias{utility_formula}
\title{Create formulas from the utility functions}
\usage{
utility_formula(x)
}
\arguments{
\item{x}{An object of class utility}
}
\value{
A list of formula expressions for the utility functions
}
\description{
Create formulas from the utility functions such that we can create correct
model matrices. The formula of each utility function contains its terms,
i.e. the utility function without the parameters.
}
\details{
Note that the formulas are created from the cleaned utility expression and
**not** the updated utility expression. This is because we are converting
dummy coded attributes to factors prior to calling \code{\link{model.matrix}}.
This ensures that dummy coded attributes are correctly returned with the
model matrix.
}
