% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.spdesign}
\alias{vcov.spdesign}
\title{Extract the variance co-variance matrix}
\usage{
\method{vcov}{spdesign}(object, ...)
}
\arguments{
\item{object}{A model object of class 'spdesign'}

\item{...}{Additional arguments passed to the function}
}
\value{
A matrix with row- and column names equal to the parameter names
}
\description{
A generic method for extracting the variance covariance matrix from a design
object
}
