% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-utility.R
\name{attribute_levels}
\alias{attribute_levels}
\title{Generic for getting the attributes and levels from the utility}
\usage{
attribute_levels(x)
}
\description{
Generic for getting the attributes and levels from the utility
}
