% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-criteria.R
\name{calculate_c_error}
\alias{calculate_c_error}
\title{C-error}
\usage{
calculate_c_error(design_vcov, p, dudx, return_all)
}
\arguments{
\item{design_vcov}{A variance-covariance matrix returned by
\code{\link{derive_vcov}} or returned by an estimation routine. The matrix
should be symmetrical and K-by-K}

\item{p}{Prior values}

\item{dudx}{A character string giving the name of the prior in the
denominator. Must be specified when optimizing for 'c-error'}

\item{return_all}{If `TRUE` return a K or K-1 vector with parameter specific error
measures. Default is `FALSE`.}
}
\value{
A vector giving the variance of the ratio for each K-1 parameter or a
single number with the sum of the variances used for optimization
}
\description{
Seeks to minimize the variance of the ratio of two parameters, for example,
willingness-to-pay.
}
