% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cor}
\alias{cor}
\title{Correlation}
\usage{
cor(x, ...)
}
\arguments{
\item{x}{A model object of class 'spdesign'}

\item{...}{Additional parameters passed to the function}
}
\value{
A matrix with correlations
}
\description{
Calculate the correlation of the design. The function gets the design from
the design object before passing it to \code{\link{cor}} from stats.
This is a wrapper around \code{\link{cor}}.
}
\details{
Note that when your design includes constants, the function will print a
 warning because the standard deviation of a constant is 0.
}
