% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc.R
\name{cycle}
\alias{cycle}
\title{Cycling of attribute levels}
\usage{
cycle(x)
}
\arguments{
\item{x}{A vector of attribute levels}
}
\value{
A cycled design candidate
}
\description{
Cycles the attribute levels to create a new design candidate. "Cycling
replaces all attribute levels in each choice situation at the time by
replacing the first level with the second level, second level with the third
etc. Since this change affects all columns, cycling can only be performed if
all attributes have exactly the same sets of feasible levels,
(e.g., where all variables are dummy coded)." (p. 253).
}
\details{
This part of the RSC algorithm is rarely invoked.
}
\references{
Hensher, D. A., Rose, J. M. & Greene, W., 2005, Applied Choice Analysis,
2nd ed., Cambridge University Press
}
