% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_vcov}
\alias{derive_vcov}
\title{Derive the variance covariance matrix of the design}
\usage{
derive_vcov(type)
}
\arguments{
\item{type}{A string indicating the model for which you wish to derive the
variance covariance matrix. Can be either "mnl" or "rpl"}
}
\description{
Derive the variance covariance matrix of the design
}
