% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{Create the experimental design}
\usage{
design(U, design_opt, candidate_set = NULL)
}
\arguments{
\item{U}{A list of utility functions}

\item{design_opt}{A list of design options}

\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no candidate
set is provided, then the full factorial subject to specified restrictions
will be used.}
}
\value{
A design object of class designer.design
}
\description{
Design takes the utility functions and design options as inputs and generates
candidate designs. Design cannot be run without first validating the utility
functions and design options. Running LINK TO VALIDATE will create a validated
environment, which is the context for evaluating the designs.
}
