% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions-and-conditions.R
\name{exclude}
\alias{exclude}
\title{Exclude rows from the candidate set}
\usage{
exclude(candidate_set, exclusions)
}
\arguments{
\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no
candidate set is provided, then the full factorial subject to specified
exclusions will be used.}

\item{exclusions}{A list of exclusions Often this list will be pulled
directly from the list of options or it is a modified list of exclusions}
}
\value{
A restricted candidate set
}
\description{
The function takes the list of exclusions and transforms them into an
expression that is then parsed and evaluated to apply the exclusions
to the supplied candidate set using standard subsetting routines.
}
