% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{extract_attrs}
\alias{extract_attrs}
\title{Extract the attributes}
\usage{
extract_attrs(V)
}
\arguments{
\item{V}{A list of utility functions}
}
\value{
A list of lists
}
\description{
Extracts the attributes, levels and interation terms
}
