% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_level_occurrence}
\alias{extract_level_occurrence}
\title{Extract the frequency of levels}
\usage{
extract_level_occurrence(string, simplify = FALSE)
}
\arguments{
\item{string}{A character string}

\item{simplify}{If TRUE return as a vector. Default is FALSE.}
}
\description{
The function extracts how many times each level of an attribute should
occur within the design when attribute level balance is not enforced.
Note that it extracts the parentheses AFTER the end of the square brackets.
Specifying round brackets without the square brackets are syntactically
invalid and therefore we want the code to fail in this case.
}
