% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-factorial.R
\name{full_factorial}
\alias{full_factorial}
\title{Generate the full factorial}
\usage{
full_factorial()
}
\description{
Wrapper function for expand.grid() to generate the full factorial design. The
full factorial will include unrealistic and completely dominated/dominating
alternatives.
}
