% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{generate_design}
\alias{generate_design}
\title{Generate the experimental design}
\usage{
generate_design(utility, opts, candidate_set = NULL)
}
\arguments{
\item{utility}{A list of utility functions}

\item{opts}{A list of design optsions}

\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no
candidate set is provided, then the full factorial subject to specified
restrictions will be used.}
}
\value{
A design list of valid designs
}
\description{
Generates the experimental design is the main function of the package. It
wraps around all other functions and calls them in the correct order to
generate a design based on the specified utility functions, options and
candidate set.
}
