% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_attribute_names}
\alias{get_attribute_names}
\title{Get the attribute names}
\usage{
get_attribute_names(object, ...)
}
\arguments{
\item{object}{A utility function of class 'utility'}

\item{...}{Additional parameters passed to the function}
}
\value{
A list with the attribute names
}
\description{
Get the attribute names
}
