% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level-occurrence.R
\name{get_level_occurrence}
\alias{get_level_occurrence}
\title{Function for getting and setting up the list of level occurrences}
\usage{
get_level_occurrence(utility, attrs, candidate_rows)
}
\arguments{
\item{utility}{A character utility with the utility expression}

\item{attrs}{A named list of attributes and levels}

\item{candidate_rows}{An integer giving the number of rows in each design
candidate, which is equal to the product of the number of alternatives,
tasks and blocks.}
}
\value{
A named list of lists where the outer list is for the attributes and
the inner list, the levels of each attribute and the number or range of times
they can occur
}
\description{
The function is called from within \code{\link{parse_utility}} with the
purpose of getting the list of attributes, levels and occurrences within the
design. This is necessary to impose attribute level balance or near attribute
level balance. This is also necessary to correctly populate the candidate
matrix used for the RSC algorithm.
}
