% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_bayesian_prior}
\alias{has_bayesian_prior}
\title{Tests whether the utility expression contains Bayesian priors}
\usage{
has_bayesian_prior(string)
}
\arguments{
\item{string}{A string or list of strings}
}
\value{
A boolean equal to `TRUE` if we have Bayesian priors
}
\description{
This is particularly useful for flow-control
}
