% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_pseudo_random}
\alias{make_pseudo_random}
\title{Make pseudo random draws}
\usage{
make_pseudo_random(n_ind, n_draws, n_dim)
}
\arguments{
\item{n_ind}{Number of individuals in your sample}

\item{n_draws}{Number of draws per respondent}

\item{n_dim}{Number of dimensions}
}
\description{
Wrapper for runif to create a common interface
}
