% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{make_standard_halton}
\alias{make_standard_halton}
\title{Wrapper for halton()}
\usage{
make_standard_halton(n_ind, n_draws, n_dim)
}
\arguments{
\item{n_ind}{Number of individuals in your sample}

\item{n_draws}{Number of draws per respondent}

\item{n_dim}{Number of dimensions}
}
\description{
Wrapper function for halton() from randtoolbox to create a common interface
}
