% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_utility_funcs.R
\name{parse_utility}
\alias{parse_utility}
\title{Parse the utility functions}
\usage{
parse_utility(V)
}
\arguments{
\item{V}{A list of utility functions}
}
\description{
This function parses the utility expressions and extracts the correct number
of attributes and levels as well as coefficients and priors.
}
\examples{
utility_funcs <- list(
  alt_1 = "beta_1[0.1] * x_1[0, 1] + beta_2[-0.1] * x_2[2, 4, 6, 8]",
  alt_2 = "beta_1 * x_1 + beta_2 * x_2"
)


}
