% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_utility}
\alias{parse_utility}
\title{Parse the utility functions}
\usage{
parse_utility(utility, tasks)
}
\arguments{
\item{utility}{A named list of utility functions}

\item{tasks}{An integer giving the number of tasks (rows) in the final design}
}
\value{
A parsed utility expression of class 'utility'. The object is a list
with the following list-elements: A cleaned utility expression, a utility
formula object, a named vector of priors, a named list of attribute levels in
the wide format, a vector of generic attribute names, and an expanded list of
attribute levels with attribute level occurences.
}
\description{
Parse the list of utility functions to extract the relevant priors, attribute
levels and other information.
}
\details{
The function is exported, but should not be used.
}
