% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prepare_priors}
\alias{prepare_priors}
\title{Prepare the list of priors}
\usage{
prepare_priors(utility, draws, R)
}
\arguments{
\item{utility}{A named list of utility functions. See the examples and the
vignette for examples of how to define these correctly for different types
of experimental designs.}

\item{draws}{The type of draws to use with Bayesian priors. No default is set
 and must be specified even if you are not creating a Bayesian design. Can be
one of "pseudo-random", "mlhs", "standard-halton", "scrambled-halton",
"standard-sobol","scrambled-sobol".}

\item{R}{An integer giving the number of draws to use. The default is 100.}
}
\value{
A list of priors
}
\description{
Prepare the list of priors
}
