% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.spdesign}
\alias{print.spdesign}
\title{A generic function for printing an 'spdesign' object}
\usage{
\method{print}{spdesign}(x, ...)
}
\arguments{
\item{x}{A model object of class 'spdesign'}

\item{...}{Additional parameters passed to the function}
}
\value{
No return value. Prints the 'spdesign' object.
}
\description{
A generic function for printing an 'spdesign' object
}
