% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_initial_header}
\alias{print_initial_header}
\title{Prints the initial header for the table of results}
\usage{
print_initial_header(efficiency_criteria, padding = 10, width = 80)
}
\arguments{
\item{efficiency_criteria}{The criteria that we optimize over}

\item{padding}{An integer specifying the padding of each column element.
Default value is 10.}

\item{width}{An integer giving the width of the horizontal rules. Default
value is 80}
}
\value{
Noting
}
\description{
The function prints the initial header for the console output and colors in
the criteria used for optimization. Effectively, the function makes multiple
calls to \code{\link{cat}}.
}
