% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{random}
\alias{random}
\title{Make a random design candidate}
\usage{
random(candidate_set, opts, J)
}
\arguments{
\item{candidate_set}{A valid candidate set}

\item{opts}{A list of design options}

\item{J}{Number of alternatives per choice task}
}
\description{
Make a random design candidate
}
