% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{random}
\alias{random}
\title{Make a random design candidate}
\usage{
random(candidate_set, opts)
}
\arguments{
\item{candidate_set}{A valid candidate set. This is only used for the
'random' and 'federov' algorithms.}

\item{opts}{A list of design options}
}
\description{
Make a random design candidate
}
