% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc.R
\name{rsc}
\alias{rsc}
\title{Make a design candidate based on the rsc algorithm}
\usage{
rsc(
  design_object,
  model,
  efficiency_criteria,
  utility,
  priors,
  didx,
  candidate_set,
  tasks,
  control
)
}
\arguments{
\item{design_object}{A list of class 'spdesign' created within the
\code{\link{generate_design}} function}

\item{model}{A character string indicating the model to optimize the design
for. Currently the only model programmed is the 'mnl' model and this is also
set as the default.}

\item{efficiency_criteria}{A character string giving the efficiency criteria
to optimize for. One of 'a-error', 'c-error', 'd-error' or 's-error'. No
default is set and argument must be specified. Optimizing for multiple
criteria is not yet implemented and will result in an error.}

\item{utility}{A named list of utility functions}

\item{priors}{A list of priors}

\item{didx}{A character string giving the name of the parameter in the
denominator. Must be specified when optimizing for 'c-error'}

\item{candidate_set}{A matrix or data frame in the "wide" format containing
all permitted combinations of attributes. The default is NULL. If no
candidate set is provided, then the full factorial subject to specified
restrictions will be used.}

\item{tasks}{An integer giving the number of tasks (rows) in the final design}

\item{control}{A list of control options}
}
\description{
Depending on the setting the function calls a combination of
\code{\link{relabel}}, \code{\link{swap}} and \code{\link{cycle}}
to create new design candidates. The code is intentionally written modular
to allow for all special cases of the algorithm.
}
