% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{rsc}
\alias{rsc}
\title{Make a design candidate based on the rsc algorithm}
\usage{
rsc(candidate_set, opts, utility)
}
\arguments{
\item{candidate_set}{A valid candidate set}

\item{opts}{A list of design options}

\item{utility}{The list of utility expressions}
}
\description{
Make a design candidate based on the rsc algorithm
}
