% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdesignR.R
\docType{package}
\name{spdesignR}
\alias{spdesignR}
\alias{spdesignR-package}
\title{spdesignR: Designing Stated Preference Experiments in R}
\description{
Contemporary software commonly used to design stated preference experiments
are expensive and the code is closed source. spdesignR is a free software
package with an easy to use interface to make flexible stated preference
experimental designs using state-of-the-art methods.
}
