% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utility.R
\name{split_utility}
\alias{split_utility}
\title{Split utility}
\usage{
split_utility(x)
}
\arguments{
\item{x}{A character string containing a single utility expression}
}
\value{
A list of character vectors where each list element contains the
parameter/value or attribute/value combinations. The first element of each
vector is always the parameter and subsequent elements are attributes or
interaction terms.
}
\description{
Takes a single utility function and splits it into pieces that can be used by
the utility parser to create the named vector of priors and the named list
of attributes and levels
}
