% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.spdesign}
\alias{summary.spdesign}
\title{Create a summary of the experimental design}
\usage{
\method{summary}{spdesign}(object, ...)
}
\arguments{
\item{object}{A model object of class 'spdesign'}

\item{...}{Additional arguments passed to the function}
}
\value{
No return value. Prints a summary of the 'spdesign' object to the
console
}
\description{
Create a summary of the experimental design
}
