% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{too_small}
\alias{too_small}
\title{Check if the design is too small}
\usage{
too_small(x, rows)
}
\arguments{
\item{x}{A list of utility expressions}

\item{rows}{The number of rows in the design}
}
\value{
A boolean equal to `TRUE` if the design is too small
}
\description{
Uses the formula of T * (J - 1) to check if the design is large enough to
identify the parameters of the utility function.
}
