% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{utility_formula}
\alias{utility_formula}
\title{Create formulas from the utility functions}
\usage{
utility_formula(x)
}
\arguments{
\item{x}{An object of class utility}
}
\value{
A list of formula expressions for the utility functions
}
\description{
Create formulas from the utility functions such that we can create correct
model matrices
}
